+++ telnyx-core/src/main/kotlin/com/telnyx/sdk/core/ClientOptions.kt
@@ -6,6 +6,7 @@ import com.fasterxml.jackson.databind.json.JsonMapper
 import com.telnyx.sdk.core.http.AsyncStreamResponse
 import com.telnyx.sdk.core.http.Headers
 import com.telnyx.sdk.core.http.HttpClient
+import com.telnyx.sdk.core.http.LoggingHttpClient
 import com.telnyx.sdk.core.http.OAuth2HttpClient
 import com.telnyx.sdk.core.http.PhantomReachableClosingHttpClient
 import com.telnyx.sdk.core.http.QueryParams
@@ -111,6 +112,14 @@ private constructor(
      * Defaults to 2.
      */
     @get:JvmName("maxRetries") val maxRetries: Int,
+    /**
+     * The level at which to log request and response information.
+     *
+     * [fromEnv] will set the level from environment variables. See [LogLevel.fromEnv].
+     *
+     * Defaults to [LogLevel.fromEnv].
+     */
+    @get:JvmName("logLevel") val logLevel: LogLevel,
     private val apiKey: String?,
     private val publicKey: String?,
     private val clientId: String?,
@@ -179,6 +188,7 @@ private constructor(
         private var responseValidation: Boolean = false
         private var timeout: Timeout = Timeout.default()
         private var maxRetries: Int = 2
+        private var logLevel: LogLevel = LogLevel.fromEnv()
         private var apiKey: String? = null
         private var publicKey: String? = null
         private var clientId: String? = null
@@ -198,6 +208,7 @@ private constructor(
             responseValidation = clientOptions.responseValidation
             timeout = clientOptions.timeout
             maxRetries = clientOptions.maxRetries
+            logLevel = clientOptions.logLevel
             apiKey = clientOptions.apiKey
             publicKey = clientOptions.publicKey
             clientId = clientOptions.clientId
@@ -325,6 +336,15 @@ private constructor(
          */
         fun maxRetries(maxRetries: Int) = apply { this.maxRetries = maxRetries }
 
+        /**
+         * The level at which to log request and response information.
+         *
+         * [fromEnv] will set the level from environment variables. See [LogLevel.fromEnv].
+         *
+         * Defaults to [LogLevel.fromEnv].
+         */
+        fun logLevel(logLevel: LogLevel) = apply { this.logLevel = logLevel }
+
         fun apiKey(apiKey: String?) = apply { this.apiKey = apiKey }
 
         /** Alias for calling [Builder.apiKey] with `apiKey.orElse(null)`. */
@@ -443,6 +463,7 @@ private constructor(
          * System properties take precedence over environment variables.
          */
         fun fromEnv() = apply {
+            logLevel(LogLevel.fromEnv())
             (System.getProperty("telnyx.baseUrl") ?: System.getenv("TELNYX_BASE_URL"))?.let {
                 baseUrl(it)
             }
@@ -526,7 +547,13 @@ private constructor(
                     .httpClient(
                         if (clientId != null && clientSecret != null) {
                             OAuth2HttpClient.builder()
-                                .httpClient(httpClient)
+                                .httpClient(
+                                    LoggingHttpClient.builder()
+                                        .httpClient(httpClient)
+                                        .clock(clock)
+                                        .level(logLevel)
+                                        .build()
+                                )
                                 .tokenUrl(
                                     (baseUrl ?: PRODUCTION_URL) +
                                         "https://api.telnyx.com/v2/oauth/token"
@@ -537,7 +564,11 @@ private constructor(
                                 .clock(clock)
                                 .build()
                         } else {
-                            httpClient
+                            LoggingHttpClient.builder()
+                                .httpClient(httpClient)
+                                .clock(clock)
+                                .level(logLevel)
+                                .build()
                         }
                     )
                     .sleeper(sleeper)
@@ -555,6 +586,7 @@ private constructor(
                 responseValidation,
                 timeout,
                 maxRetries,
+                logLevel,
                 apiKey,
                 publicKey,
                 clientId,